requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
python-docx
sentence-transformers
transformers
torch
langdetect
